refactor(hush-cli): extract util from pkg_cli
Move shared helpers urlencoding_simple, truncate_with_ellipsis,
format_number, tempdir_for_download (and the private HEX_UPPER const)
into pkg_cli/util.rs, widened fn -> pub(super) for cross-module use.
Doc comments moved verbatim with their functions.
