+++ crates/services/hush-cli/src/pkg_cli/mod.rs
@@ -19,8 +19,10 @@ const PLUGIN_MANIFEST_FILENAME: &str = "clawdstrike.plugin.toml";
 const MAX_REGISTRY_DOWNLOAD_BYTES: u64 = 100 * 1024 * 1024;
 
 mod command;
+mod util;
 
 pub use command::{CliPkgType, OrgCommands, PkgCommands, TrustedPublisherCommands};
+use util::{format_number, tempdir_for_download, truncate_with_ellipsis, urlencoding_simple};
 
 // ---------------------------------------------------------------------------
 // Dispatcher
@@ -1656,13 +1658,6 @@ fn verify_install_trust(
     }
 }
 
-fn tempdir_for_download() -> std::io::Result<PathBuf> {
-    let nonce: u64 = rand::Rng::random(&mut rand::rng());
-    let dir = std::env::temp_dir().join(format!("clawdstrike_dl_{nonce:x}"));
-    std::fs::create_dir_all(&dir)?;
-    Ok(dir)
-}
-
 struct CallerAuthHeaders {
     key_hex: String,
     sig_hex: String,
@@ -2855,34 +2850,6 @@ fn cmd_pkg_search(
     ExitCode::Ok
 }
 
-/// Minimal percent-encoding for query parameters (avoids pulling in another dep).
-fn urlencoding_simple(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    for b in s.bytes() {
-        match b {
-            b'A'..=b'Z' | b'a'..=b'z' | b'0'..=b'9' | b'-' | b'_' | b'.' | b'~' => {
-                out.push(b as char);
-            }
-            _ => {
-                out.push('%');
-                out.push(char::from(HEX_UPPER[(b >> 4) as usize]));
-                out.push(char::from(HEX_UPPER[(b & 0x0f) as usize]));
-            }
-        }
-    }
-    out
-}
-
-fn truncate_with_ellipsis(input: &str, max_chars: usize) -> String {
-    if input.chars().count() <= max_chars {
-        return input.to_string();
-    }
-    let truncated: String = input.chars().take(max_chars).collect();
-    format!("{truncated}...")
-}
-
-const HEX_UPPER: [u8; 16] = *b"0123456789ABCDEF";
-
 // ---------------------------------------------------------------------------
 // pkg audit
 // ---------------------------------------------------------------------------
@@ -3084,19 +3051,6 @@ fn cmd_pkg_stats(
     }
 
     ExitCode::Ok
-}
-
-/// Format a number with comma separators (e.g. 1234 -> "1,234").
-fn format_number(n: u64) -> String {
-    let s = n.to_string();
-    let mut result = String::with_capacity(s.len() + s.len() / 3);
-    for (i, c) in s.chars().rev().enumerate() {
-        if i > 0 && i % 3 == 0 {
-            result.push(',');
-        }
-        result.push(c);
-    }
-    result.chars().rev().collect()
 }
 
 // ---------------------------------------------------------------------------
